+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Fixed system resources components not being updated
 - Changed get system resources methods to recover only the last resource
-- Updated various components to a new responsive and self loading model
+- Updated all data dependent components to a new responsive and self loading model (load data independently from layout)
 - Created the database properties dashboard (From the environment view menu)
 
 # 0.4.0